[DEV-144] fix: 모킹 용으로 분리한 route handler 디렉토리 기존대로 복구
