+++ src/app/api/v1/goals/progress/route.ts
